requirements.txt
python-dotenv>=1.0.1
streamlit>=1.32.0
sentence-transformers>=2.6.0

langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.0.3
langchain-unstructured>=0.1.0
langchain-chroma>=0.1.0
langchain-groq>=0.1.4

unstructured[pdf]>=0.12.5
chromadb>=0.4.24
